requirements.txt
dask[distributed]
rioxarray
scipy
black[jupyter]
bokeh
cartopy
contextily
flake8
geopandas
jupyterlab
matplotlib_scalebar
mapclassify
pooch
pylint
